pyproject.toml
[tool.poetry]
name = "mmfm"
version = "0.1.0"
description = ""
authors = ["dbb"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
tqdm = "^4.66.4"
pillow = "^10.3.0"
google-cloud = "^0.34.0"
google-auth = "^2.30.0"
google-cloud-storage = "^2.17.0"
pyyaml = "^6.0.1"
openpyxl = "^3.1.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
